Last part in windows support, director on windows.
This patch adds the things needed to cross-compile the director using
MINGW on Linux. It adds the build infrastructure for building the
database backend infrastructure (catalog) and the actual director and
dbcheck for windows. It also adds the bsmtp binary in tools and the
changes needed to the installer to install all new stuff.

It does however not supply the database (postgresql or sqlite3) you
need to install those before running the bareos-installer.

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>