Adds support for different blocksizes.
Min- and maxblocksize are now a property of the
pool and are promoted to tapes in that pool.

Therefore, the media and pool database tables were
adapted to hold min and maxblocksize.

In order to be able to read the tape label always,
the label is always read and written with the
DEFAULT_BLOCK_SIZE = 63k/64512 bytes

For backward compatibility, the blocksize
DEFAULT_BLOCK_SIZE is used if the blocksize
for the medium is zero in the database table.

The device blocksizes are set to the values
of the mounted volume after it was mounted, so
all blocks being written after the label block
are written with the blocksizes configured in the
medium/pool.

Moving a tape from one pool to another through
the scratch pool mechanism is supported,
as then the pool's properties are promoted
to the medium.

This new feature makes it possible to move from
one blocksize to a new one during normal operation
while being able to restore the old backups at any
time.

For disaster recovery (DR) using bscan etc. you need to
set the blocksize used during write of the tape for
restore.

Fixes #267: make blocksize a property of the pool