check cfg and db conn before startup of director
We now use the possibilty to check the configuration
and database connection on startup of the director before
we really startup the daemon so that we get direct
feedback if something is wrong.

Before, we did not get any feedback because the director
forked so problems only were visible in the logfile.

Without systemd, we get the output directly, with systemd we
need to call "systemctl status bareos-dir.service"

Fixes #286: Alter the startscripts to run the daemons with -t before starting them

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>