Abstract the accurate payload storage from the datatype used.
Up until now the accurate payload was always stored in a htable
(memory hash table) but that will not scale. We want to support multiple
storage classes for the accurate payload (e.g. the actual accurate data
accessed via the filename as key.) Things like Symas LMDB is a good
candidate as a key/value store. This patch implements the accurate data
using a class based abstraction that allows multiple implementations of
the storage. Currently it supports the old HTABLE and LMDB. The
prototype has some code that is not needed at this point as accurate
data is written only ones and after that never updated so we don't
strictly need an update method which is implemented for LMDB and is
unneeded for HTABLE as there you get internal memory you can manipulate.
The seen logic is moved into a simple bitmap that is part of the higher
level B_ACCURATE class so we don't need to update the actual accurate
payload when we only want to register that the item has been vistited.
Otherwise for things like LMDB we keep on updating the data in the LMDB
which is just to much overhead which we can easily work around with this
bitmap.