handle unset parameter
some scripts (eg. package postinstall script) include bareos-config-lib.sh with settings,
that exit if a variable is undefined.
To prevent problems, function parameter are now used as "${1-}", so that they are never undefined.
Also fixes a bug, when using a database password that contains "/".
"/" is now a allowed character, but "#" is forbidden.

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>